# docs/specs/Swagger-V1.yaml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: API
  description: Fanlink API
  termsOfService: ''
  contact:
    name: ''
  license:
    name: ''
host: localhost
basePath: "/"
schemes:
- http
- https
...
consumes:
- application/json
produces:
- application/json
paths:
...
  "/people":
    get:
      parameters:
      - in: query
        name: username_filter
        required: true
        type: string
        description: A username or username fragment to filter on.
        example: Admin
      - in: query
        name: email_filter
        required: true
        type: string
        description: An email or email fragment to filter on.
        example: admin@example.com
      responses:
        '200':
          schema:
            type: object
            properties:
              people:
                type: array
                items:
                  oneOf:
                  - "$ref": "#/definitions/person_response"
                  - "$ref": "#/definitions/person_private_response"
                  discriminator:
                    propertyName: type
                    mapping:
                      Public Person App Response: "#/definitions/person_response"
                      Private Person App Response: "#/definitions/person_private_response"
            required:
            - people
        '401':
          description: User is not authorized to access this endpoint.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '500':
          description: Internal Server Error. Server threw an unrecoverable error.
            Create a ticket with any form fields you we're trying to send, the URL,
            API version number and any steps you took so that it can be replicated.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
      description: This is used to get a list of people.
    post:
      parameters:
      - name: input
        in: body
        required: true
        schema:
          type: object
          properties:
            product:
              type: string
              description: Internal name of the product.
            person:
              type: object
              properties:
                username:
                  type: string
                  description: Username. This needs to be unique within product scope.
                  example: test_robot
                email?:
                  type: string
                  description: Email address (required unless using FB auth token).
                  example: hello@example.com
                name?:
                  type: string
                  description: User's name.
                  example: Test McTesterson
                facebook_auth_token!:
                  type: string
                  description: Auth token from Facebook.
                picture?:
                  type: string
                  format: binary
                  description: Profile picture, this should be `image/gif`, `image/png`,
                    or `image/jpeg`.
                gender:
                  type: integer
                  format: int32
                  description: 'Gender. Valid options: unspecified (default), male,
                    female.'
                  example: 'TODO: Example'
                birthdate?:
                  type: string
                  format: date
                  description: Birthdate in format 'YYYY-MM-DD'.
                  example: '1970-08-02'
                city?:
                  type: string
                  description: Person's supplied city.
                  example: Las Vegas
                country_code?:
                  type: string
                  description: Alpha2 code (two letters) from ISO 3166 list.
                  example: US
              required:
              - username
              - email?
              - name?
              - facebook_auth_token!
              - picture?
              - gender
              - birthdate?
              - city?
              - country_code?
          required:
          - product
          - person
      responses:
        '200':
          schema:
            type: object
            properties:
              person:
                type: object
                properties:
                  type:
                    "$ref": "#/definitions/person_private_response"
                required:
                - type
            required:
            - person
        '401':
          description: User is not authorized to access this endpoint.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '404':
          description: The record was not found.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '422':
          description: One or more fields were invalid. Check response for reasons.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '429':
          description: Not enough time since last submission of this action type or
            duplicate action type, person, identifier combination.
          schema:
            type: string
        '500':
          description: Internal Server Error. Server threw an unrecoverable error.
            Create a ticket with any form fields you we're trying to send, the URL,
            API version number and any steps you took so that it can be replicated.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
      description: |-
        This is used to create a new person.
        If the account creation is successful, they will be logged in and we will send an onboarding email (if we have an email address for them).
  "/people/{id}":
    get:
      parameters:
      - in: path
        name: id
        required: true
        type: integer
        format: int32
      responses:
        '200':
          schema:
            type: object
            properties:
              person:
                type: object
                properties:
                  type:
                    oneOf:
                    - "$ref": "#/definitions/person_response"
                    - "$ref": "#/definitions/person_private_response"
                    discriminator:
                      propertyName: type
                      mapping:
                        Public Person App Response: "#/definitions/person_response"
                        Private Person App Response: "#/definitions/person_private_response"
                required:
                - type
            required:
            - person
        '401':
          description: User is not authorized to access this endpoint.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '404':
          description: The record was not found.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '500':
          description: Internal Server Error. Server threw an unrecoverable error.
            Create a ticket with any form fields you we're trying to send, the URL,
            API version number and any steps you took so that it can be replicated.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
      description: Returns the person for the supplied ID.
    put:
      parameters:
      - in: path
        name: id
        required: true
        type: integer
        format: int32
      - name: input
        in: body
        required: true
        schema:
          type: object
          properties:
            recommended:
              type: boolean
              description: Whether this is a recommended persion. (Admin or product
                account only)
            person:
              type: object
              properties:
                username:
                  type: string
                  description: Username. This needs to be unique within product scope.
                  example: test_robot
                email?:
                  type: string
                  description: Email address (required unless using FB auth token).
                  example: hello@example.com
                name?:
                  type: string
                  description: User's name.
                  example: Test McTesterson
                picture?:
                  type: string
                  format: binary
                  description: Profile picture, this should be `image/gif`, `image/png`,
                    or `image/jpeg`.
                gender:
                  type: integer
                  format: int32
                  description: 'Gender. Valid options: unspecified (default), male,
                    female.'
                  example: 'TODO: Example'
                birthdate?:
                  type: string
                  format: date
                  description: Birthdate in format 'YYYY-MM-DD'.
                  example: '1970-08-02'
                city?:
                  type: string
                  description: Person's supplied city.
                  example: Las Vegas
                country_code?:
                  type: string
                  description: Alpha2 code (two letters) from ISO 3166 list.
                  example: US
              required:
              - username
              - email?
              - name?
              - picture?
              - gender
              - birthdate?
              - city?
              - country_code?
          required:
          - recommended
          - person
      responses:
        '200':
          schema:
            type: object
            properties:
              person:
                type: object
                properties:
                  type:
                    "$ref": "#/definitions/person_private_response"
                required:
                - type
            required:
            - person
        '401':
          description: User is not authorized to access this endpoint.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '404':
          description: The record was not found.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '422':
          description: One or more fields were invalid. Check response for reasons.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
        '500':
          description: Internal Server Error. Server threw an unrecoverable error.
            Create a ticket with any form fields you we're trying to send, the URL,
            API version number and any steps you took so that it can be replicated.
          schema:
            type: object
            properties:
              errors:
                type: object
                properties:
                  base:
                    type: array
                    items:
                      type: string
                required:
                - base
            required:
            - errors
...
definitions:
...
  level_response:
    type: object
    properties:
      id:
        type: string
      internal_name:
        type: string
      points:
        type: integer
        format: int32
      picture_url:
        type: string
      description:
        type: string
      name:
        type: string
    required: []
...
  person_private_response:
    type: object
    properties:
      id:
        type: string
      username:
        type: string
      email:
        type: string
      name:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
      facebookid:
        type: string
      facebook_picture_url:
        type: string
      picture_url:
        type: string
      num_followers:
        type: integer
        format: int32
      num_followings:
        type: integer
        format: int32
      following_id:
        type: integer
        format: int32
      level:
        "$ref": "#/definitions/level_response"
      badge_points:
        type: integer
        format: int32
      do_not_message_me:
        type: boolean
      pin_messages_from:
        type: boolean
      auto_follow:
        type: boolean
      role:
        type: string
        enum:
        - normal
        - staff
        - admin
        - super_admin
      product_account:
        type: boolean
      chat_banned:
        type: boolean
      designation:
        type: string
      recommended:
        type: boolean
      gender:
        type: string
        enum:
        - male
        - female
        - unspecified
      birthdate:
        type: string
      city:
        type: string
      country_code:
        type: string
      biography:
        type: string
      product:
        type: object
        properties:
          internal_name:
            type: string
          id:
            type: integer
            format: int32
          name:
            type: string
        required: []
    required: []
  person_response:
    type: object
    properties:
      id:
        type: string
      username:
        type: string
      name:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
      facebookid:
        type: string
      facebook_picture_url:
        type: string
      picture_url:
        type: string
      num_followers:
        type: integer
        format: int32
      num_followings:
        type: integer
        format: int32
      following_id:
        type: integer
        format: int32
      level:
        "$ref": "#/definitions/level_response"
      badge_points:
        type: integer
        format: int32
      do_not_message_me:
        type: boolean
      pin_messages_from:
        type: boolean
      auto_follow:
        type: boolean
      role:
        type: string
        enum:
        - normal
        - staff
        - admin
        - super_admin
      product_account:
        type: boolean
      chat_banned:
        type: boolean
      designation:
        type: string
      recommended:
        type: boolean
      gender:
        type: string
        enum:
        - male
        - female
        - unspecified
      birthdate:
        type: string
      city:
        type: string
      country_code:
        type: string
      biography:
        type: string
    required: []
...